Move events enum outside api.ts
